Fixed Assign and Updated CRUD with "Are you sure?" DELETE pop-up
Fixed all of Assign for UPDATE and added "Are you sure?" DELETE pop-up in all of CRUD code.

+++ project/CRUD/announcements.php
@@ -138,7 +138,7 @@ if (isset($_GET['delete_id'])) {
                     <td><?= $course_name ?></td>
                     <?php if (!isStudent()) {
                         echo '<td><a href="?page=announcements&update_view=true&update_id=' . $id . '">Update</a></td>';
-                        echo '<td><a href="?page=announcements&delete_view=true&delete_id=' . $id . '">Delete</a></td>';
+                        echo "<td><a href='?page=announcements&delete_view=true&delete_id=" . $id . "' onclick='return confirm(&quot;Are you sure you want to delete? &quot;)'>Delete</a></td>";
                     } ?>
                 </tr>
             <?php
