+++ project/CRUD/users.php
@@ -323,7 +323,7 @@ if (isset($_GET['delete_id'])) {
                     <td><?= $role_name ?></td>
                     <?php if (isAdmin()) {
                         echo '<td><a href="?page=users&update_view=true&update_id=' . $user_id . '">Update</a></td>';
-                        echo '<td><a href="?page=users&delete_view=true&delete_id=' . $user_id . '&role_id=' . $role_id . '">Delete</a></td>';
+                        echo "<td><a href='?page=users&delete_view=true&delete_id=" . $user_id . "&role_id=" . $role_id . "' onclick='return confirm(&quot;Are you sure you want to delete? &quot;)'>Delete</a></td>";
                     } ?>
                 </tr>
             <?php
