+++ project/assigns/assign-professors.php
@@ -33,12 +33,14 @@ if (isset($_POST['assign'])) {
     //        array_push($errors, "Please select a user");
     //    } else {
     $user_id = mysqli_real_escape_string($conn, $_POST['user_id']);
+	$_GET['user_id'] = $user_id;
     //    }
 
     //    if (empty($_POST['course_id'])) {
     //        array_push($errors, "Please select a course");
     //    } else {
     $course_id = mysqli_real_escape_string($conn, $_POST['course_id']);
+	$_GET['course_id'] = $course_id;
     //    }
 
     $query = "SELECT * FROM user_course_section WHERE user_id = '$user_id'";
@@ -54,14 +56,14 @@ if (isset($_POST['assign'])) {
 
     if (count($errors) == 0) {
 
-        $add = "INSERT INTO user_course_section (user_id, course_id) VALUES('$user_id', '$course_id')";
+        $add = "INSERT INTO user_course_section (user_id, course_id, section_id) VALUES('$user_id', '$course_id', null)";
 
         if (mysqli_query($conn, $add)) {
-            array_push($success, "Professor Assigned Successfully");
+            array_push($success, "Professor successfully assigned to this course.");
             // clear variables
             $user_id = $course_id = $section_id = $user_id_selected = $course_id_selected = "";
         } else {
-            array_push($errors, "Adding Error: " . mysqli_error($conn));
+            array_push($errors, "Could not INSERT error: " . mysqli_error($conn));
         }
     }
 }
@@ -71,41 +73,36 @@ if (isset($_POST['assign'])) {
  ********************************************************/
 
 if (isset($_POST['update'])) {
-
+	
     $user_id = mysqli_real_escape_string($conn, $_GET['user_id']);
 
     //    if (empty($_POST['course_id'])) {
     //        array_push($errors, "Please select a course");
     //    } else {
     $course_id = mysqli_real_escape_string($conn, $_POST['course_id']);
+	$_GET['course_id'] = $course_id;
     //    }
 
     $query = "SELECT * FROM user_course_section WHERE user_id = '$user_id'";
     $check = mysqli_query($conn, $query);
 
-    // foreach ($check as $row) {
-    //     $check_course_id = $row['course_id'];
-    //     $check_section_id = $row['section_id'];
-
-    //     if ($check_course_id == $course_id) {
-    //         array_push($errors, "User already assigned to this course");
-    //         if ($check_section_id == $section_id) {
-    //             array_push($errors, "User already assigned to this section");
-    //         }
-    //     }
-    // }
-
+    foreach ($check as $row) {
+        $check_course_id = $row['course_id'];
+		if ($check_course_id == $course_id) {
+				array_push($errors, "Professor is already assigned to this course.");
+			}
+	}
+    
     if (count($errors) == 0) {
 
-        $update = "UPDATE user_course_section set user_id = '$user_id', course_id = '$course_id'
-            WHERE user_id ='$user_id' AND course_id = '$course_id'";
-
+        $update = "UPDATE user_course_section set course_id = '$course_id', section_id = null WHERE user_id ='$user_id'";
+		
         if (mysqli_query($conn, $update)) {
-            array_push($success, "Updated Successfully");
+            array_push($success, "Updated successfully.");
             // clear variables
             $user_id = $course_id = $section_id = $user_id_selected = $course_id_selected = "";
         } else {
-            array_push($errors, "Error: " . mysqli_error($conn));
+            array_push($errors, "Could not UPDATE error: " . mysqli_error($conn));
         }
     }
 }
@@ -119,21 +116,20 @@ if (isset($_GET['delete_view'])) {
     $user_id = mysqli_real_escape_string($conn, $_GET['user_id']);
     $course_id = mysqli_real_escape_string($conn, $_GET['course_id']);
 
-    $delete = "DELETE FROM user_course_section
-                WHERE user_id='$user_id' AND course_id='$course_id'";
+    $delete = "DELETE FROM user_course_section WHERE user_id='$user_id' AND course_id='$course_id'";
 
     if (mysqli_query($conn, $delete)) {
-        array_push($success, "Delete successful");
+        array_push($success, "Delete successful.");
     } else {
-        array_push($errors, "Delete error: " . mysqli_error($conn));
+        array_push($errors, "Could not DELETE error: " . mysqli_error($conn));
     }
 }
 
-if (isset($_POST["course_id"])) {
-
-    $user_id_selected = $_POST["user_id"];
-    $course_id_selected = $_POST["course_id"];
-}
+//if (isset($_POST["course_id"])) {
+//
+//    $user_id_selected = $_POST["user_id"];
+//    $course_id_selected = $_POST["course_id"];
+//}
 
 ?>
 
@@ -182,8 +178,8 @@ Always visible and shows delete error if delete_view is set true -->
                     <td><?php echo $professor_id ?></td>
                     <td><?php echo $first_name . " " . $last_name ?></td>
                     <td><?php echo $course_name ?></td>
-                    <td><a href="?page=assign-professors&update_view=true&user_id=<?= $user_id ?>&course_id=<?= $course_id ?>">Update</a></td>
-                    <td><a href="?page=assign-professors&delete_view=true&user_id=<?= $user_id ?>&course_id=<?= $course_id ?>">Delete</a></td>
+                    <!-- <td><a href="?page=assign-professors&update_view=true&user_id=<?= $user_id ?>&course_id=<?= $course_id ?>">Change Course</a></td> -->
+                    <td><a href="?page=assign-professors&delete_view=true&user_id=<?= $user_id ?>&course_id=<?= $course_id ?>" onclick="return confirm('Are you sure you want to delete?')">Delete Course</a></td>
                 </tr>
             <?php
             }
@@ -275,6 +271,7 @@ Always visible and shows delete error if delete_view is set true -->
         JOIN user_course_section as ucs ON ucs.user_id = u.user_id
         JOIN course as c ON c.course_id = ucs.course_id
         WHERE u.user_id='$user_id' AND c.course_id = '$course_id'";
+		
         $results = mysqli_query($conn, $query);
 
         foreach ($results as $row) {
@@ -293,38 +290,14 @@ Always visible and shows delete error if delete_view is set true -->
                 ?>
 
                 <div class="form-input">
-                    <label>Professor</label>
+                    <label>Professor: </label>
                     <span><b><?= $professor_name ?></b></span>
-
-                    <!-- <div class="scroll-list">
-                        <select name=user_id>
-                            <option value="" selected hidden>Choose a User</option>
-                            <?php
-                            // $query = "SELECT * FROM users as u
-                            //             JOIN professor as st ON st.user_id = u.user_id
-                            //             WHERE role_id != 1";
-                            // $users = mysqli_query($conn, $query);
-                            // foreach ($users as $user) {
-                            //     $user_id = $user['user_id'];
-                            //     $first_name = $user['first_name'];
-                            //     $last_name = $user['last_name'];
-                            //     if ($update_user_id == $user_id) {
-                            //         echo "<option value='$user_id' selected>$first_name $last_name</option>";
-                            //     } else {
-                            //         echo "<option value='$user_id'>$first_name $last_name</option>";
-                            //     }
-                            // }
-                            ?>
-                        </select>
-                    </div> -->
-
-
                 </div>
 
                 <div class="form-input">
                     <p>Courses</p>
                     <div class="scroll-list">
-                        <select name=course_id>
+                        <select name="course_id" id="course_id">
                             <option value="" selected hidden>Choose a Course</option>
                             <?php
                             $query = "SELECT * FROM course";
@@ -344,7 +317,7 @@ Always visible and shows delete error if delete_view is set true -->
                 </div>
 
                 <div class="form-submit">
-                    <input type="submit" name="Submit" value="Update">
+                    <input type="submit" name="update" value="Update">
                 </div>
             </form>
         </div>
