+++ project/assigns/assign-students.php
@@ -62,24 +62,24 @@ if (isset($_POST['assign'])) {
         $check_course_id = $row['course_id'];
         $check_section_id = $row['section_id'];
 
-        if ($check_course_id == $course_id) {
-            array_push($errors, "Student is already assigned to this course.");
-            if ($check_section_id == $section_id) {
-                array_push($errors, "Student is already assigned to this section.");
-            }
-        }
-    }
+		if ($check_course_id == $course_id) {
+			array_push($errors, "This student is already assigned to this course.");
+			if ($check_section_id == $section_id) {
+				array_push($errors, "This student is already assigned to this section.");
+			}
+		}
+	}
 
     if (count($errors) == 0) {
 
         $add = "INSERT INTO user_course_section (user_id, course_id, section_id) VALUES('$user_id', '$course_id', '$section_id')";
 
         if (mysqli_query($conn, $add)) {
-            array_push($success, "Student Assigned Successfully");
+            array_push($success, "Student has been assigned to course successfully");
             // clear variables
             $user_id = $course_id = $section_id = $user_id_selected = $course_id_selected = "";
         } else {
-            array_push($errors, "Adding Error: " . mysqli_error($conn));
+            array_push($errors, "Could not INSERT error: " . mysqli_error($conn));
         }
     }
 }
@@ -95,42 +95,41 @@ if (isset($_POST['update'])) {
     //    if (empty($_POST['course_id'])) {
     //        array_push($errors, "Please select a course");
     //    } else {
-    $course_id = mysqli_real_escape_string($conn, $_POST['course_id']);
+    $course_id = mysqli_real_escape_string($conn, $_GET['course_id']);
     //    }
 
     //    if (empty($_POST['section_id'])) {
     //        array_push($errors, "Please select a section");
     //    } else {
     $section_id = mysqli_real_escape_string($conn, $_POST['section_id']);
-    //    }
-
+	$_GET['section_id'] = $section_id;
+    // }
 
     $query = "SELECT * FROM user_course_section WHERE user_id = '$user_id'";
     $check = mysqli_query($conn, $query);
 
-    // foreach ($check as $row) {
-    //     $check_course_id = $row['course_id'];
-    //     $check_section_id = $row['section_id'];
-
-    //     if ($check_course_id == $course_id) {
-    //         array_push($errors, "User already assigned to this course");
-    //         if ($check_section_id == $section_id) {
-    //             array_push($errors, "User already assigned to this section");
-    //         }
-    //     }
-    // }
-
+    foreach ($check as $row) {
+		$check_course_id = $row['course_id'];
+		$check_section_id = $row['section_id'];
+	
+		// Check to see if the Student is already assigned the same course and/or section
+		if ($check_course_id == $course_id) {
+			if ($check_section_id == $section_id) {
+				array_push($errors, "Student is already assigned to this course and section.");
+			}
+		}
+	}
+	
     if (count($errors) == 0) {
 
-        $update = "UPDATE user_course_section set user_id = '$user_id', course_id = '$course_id', section_id = '$section_id'
-            WHERE user_id ='$user_id' AND section_id = '$section_id'";
+        $update = "UPDATE user_course_section set section_id = '$section_id' WHERE user_id ='$user_id' AND course_id = '$course_id'";
 
         if (mysqli_query($conn, $update)) {
-            array_push($success, "Updated Successfully");
+            array_push($success, "Updated Successfully.");
             // clear variables
             $user_id = $course_id = $section_id = $user_id_selected = $course_id_selected = "";
         } else {
-            array_push($errors, "Error: " . mysqli_error($conn));
+            array_push($errors, "Could not UPDATE error: " . mysqli_error($conn));
         }
     }
 }
@@ -145,13 +144,12 @@ if (isset($_GET['delete_view'])) {
     $course_id = mysqli_real_escape_string($conn, $_GET['course_id']);
     $section_id = mysqli_real_escape_string($conn, $_GET['section_id']);
 
-    $delete = "DELETE FROM user_course_section
-                WHERE user_id='$user_id' AND course_id='$course_id' AND section_id='$section_id'";
+    $delete = "DELETE FROM user_course_section WHERE user_id='$user_id' AND course_id='$course_id' AND section_id='$section_id'";
 
     if (mysqli_query($conn, $delete)) {
-        array_push($success, "Delete successful");
+        array_push($success, "Delete successful.");
     } else {
-        array_push($errors, "Delete error: " . mysqli_error($conn));
+        array_push($errors, "Could not DELETE error: " . mysqli_error($conn));
     }
 }
 
@@ -214,8 +212,8 @@ Always visible and shows delete error if delete_view is set true -->
                     <td><?php echo $first_name . " " . $last_name ?></td>
                     <td><?php echo $course_name ?></td>
                     <td><?php echo $section_name ?></td>
-                    <td><a href="?page=assign-students&update_view=true&user_id=<?= $user_id ?>&course_id=<?= $course_id ?>&section_id=<?= $section_id ?>">Update</a></td>
-                    <td><a href="?page=assign-students&delete_view=true&user_id=<?= $user_id ?>&course_id=<?= $course_id ?>&section_id=<?= $section_id ?>">Delete</a></td>
+                    <td><a href="?page=assign-students&update_view=true&user_id=<?= $user_id ?>&course_id=<?= $course_id ?>&section_id=<?= $section_id ?>">Change Section</a></td>
+                    <td><a href="?page=assign-students&delete_view=true&user_id=<?= $user_id ?>&course_id=<?= $course_id ?>&section_id=<?= $section_id ?>" onclick="return confirm('Are you sure you want to delete?')">Delete Course</a></td>
                 </tr>
             <?php
             }
@@ -339,12 +337,14 @@ Always visible and shows delete error if delete_view is set true -->
         JOIN course as c ON c.course_id = ucs.course_id
         JOIN section as s ON s.section_id = ucs.section_id
         WHERE u.user_id='$user_id' AND c.course_id = '$course_id' AND s.section_id = '$section_id'";
+		
         $results = mysqli_query($conn, $query);
-
+		
         foreach ($results as $row) {
             $student_name = $row['first_name'] . " " . $row['last_name'];
-            $update_user_id = $row['user_id'];
-            $update_course_id = $row['course_id'];
+            $user_id = $row['user_id'];
+            $course_id = $row['course_id'];
+			$course_name = $row['course_name'];
             $update_section_name = $row['section_name'];
         }
         ?>
@@ -358,66 +358,32 @@ Always visible and shows delete error if delete_view is set true -->
                 ?>
 
                 <div class="form-input">
-                    <label>Student</label>
+                    <label>Student: </label>
                     <span><b><?= $student_name ?></b></span>
-
-                    <!-- <div class="scroll-list">
-                        <select name=user_id>
-                            <option value="" selected hidden>Choose a User</option>
-                            <?php
-                            // $query = "SELECT * FROM users as u
-                            //             JOIN student as st ON st.user_id = u.user_id
-                            //             WHERE role_id != 1";
-                            // $users = mysqli_query($conn, $query);
-                            // foreach ($users as $user) {
-                            //     $user_id = $user['user_id'];
-                            //     $first_name = $user['first_name'];
-                            //     $last_name = $user['last_name'];
-                            //     if ($update_user_id == $user_id) {
-                            //         echo "<option value='$user_id' selected>$first_name $last_name</option>";
-                            //     } else {
-                            //         echo "<option value='$user_id'>$first_name $last_name</option>";
-                            //     }
-                            // }
-                            ?>
-                        </select>
-                    </div> -->
-
-
                 </div>
 
                 <div class="form-input">
-                    <p>Courses</p>
+                    <p>Course:</p>
                     <div class="scroll-list">
-                        <select name="course_id" disabled>
-                            <option value="" selected hidden>Choose a Course</option>
+                        <select name="course_id" id="course_id" disabled>
                             <?php
-                            $query = "SELECT * FROM course";
-                            $courses = mysqli_query($conn, $query);
-                            foreach ($courses as $row) {
-                                $course_id = $row['course_id'];
-                                $course_name = $row['course_name'];
-                                if ($update_course_id == $course_id) {
-                                    echo "<option value='$course_id' selected>$course_name</option>";
-                                } else {
-                                    echo "<option value='$course_id'>$course_name</option>";
-                                }
-                            }
+								echo "<option value='$course_id' selected>$course_name</option>";
                             ?>
                         </select>
                     </div>
                 </div>
 
                 <div class="form-input">
-                    <p>Sections</p>
+                    <p>Course Sections:</p>
                     <div class="scroll-list">
-                        <select name=section_id>
-                            <option value="" selected hidden>Choose a Section</option>
+                        <select name="section_id" id="section_id">
+                            <option value="" selected hidden>Choose a section:</option>
                             <?php
-
+							
+							// Get limited section names based on course_id
                             $query = "SELECT * FROM section as s
                                         JOIN course as c ON c.course_id = s.course_id
-										WHERE c.course_id = '$update_course_id'";
+										WHERE c.course_id = '$course_id'";
 
                             $sections = mysqli_query($conn, $query);
                             foreach ($sections as $row) {
